Refactor - made card specific effects to attack random opponent shared effects
