+++ Mage.Sets/src/mage/sets/commander/RuhanOfTheFomori.java
@@ -27,27 +27,14 @@
  */
 package mage.sets.commander;
 
-import java.util.ArrayList;
-import java.util.List;
-import java.util.Random;
 import java.util.UUID;
 import mage.MageInt;
-import mage.abilities.Ability;
 import mage.abilities.common.BeginningOfCombatTriggeredAbility;
-import mage.abilities.effects.ContinuousEffect;
-import mage.abilities.effects.OneShotEffect;
-import mage.abilities.effects.RequirementEffect;
+import mage.abilities.effects.common.combat.AttackIfAbleTargetRandomOpponentSourceEffect;
 import mage.cards.CardImpl;
 import mage.constants.CardType;
-import mage.constants.Duration;
-import mage.constants.Outcome;
 import mage.constants.Rarity;
 import mage.constants.TargetController;
-import mage.game.Game;
-import mage.game.permanent.Permanent;
-import mage.players.Player;
-import mage.target.targetpointer.FixedTarget;
-import mage.util.RandomUtil;
 
 /**
  *
@@ -66,7 +53,7 @@ public class RuhanOfTheFomori extends CardImpl {
         this.toughness = new MageInt(7);
 
         // At the beginning of combat on your turn, choose an opponent at random. Ruhan of the Fomori attacks that player this combat if able.
-        this.addAbility(new BeginningOfCombatTriggeredAbility(new RuhanOfTheFomoriEffect(), TargetController.YOU, false));
+        this.addAbility(new BeginningOfCombatTriggeredAbility(new AttackIfAbleTargetRandomOpponentSourceEffect(), TargetController.YOU, false));
     }
 
     public RuhanOfTheFomori(final RuhanOfTheFomori card) {
@@ -79,77 +66,3 @@ public class RuhanOfTheFomori extends CardImpl {
     }
 }
 
-class RuhanOfTheFomoriEffect extends OneShotEffect {
-
-    public RuhanOfTheFomoriEffect() {
-        super(Outcome.Benefit);
-        this.staticText = "choose an opponent at random. {this} attacks that player this combat if able";
-    }
-
-    public RuhanOfTheFomoriEffect(final RuhanOfTheFomoriEffect effect) {
-        super(effect);
-    }
-
-    @Override
-    public RuhanOfTheFomoriEffect copy() {
-        return new RuhanOfTheFomoriEffect(this);
-    }
-
-    @Override
-    public boolean apply(Game game, Ability source) {
-        Player controller = game.getPlayer(source.getControllerId());
-        if (controller != null) {
-            List<UUID> opponents = new ArrayList<>();
-            opponents.addAll(game.getOpponents(controller.getId()));
-            Player opponent = game.getPlayer(opponents.get(RandomUtil.nextInt(opponents.size())));
-            if (opponent != null) {
-                ContinuousEffect effect = new AttacksIfAbleTargetPlayerSourceEffect();
-                effect.setTargetPointer(new FixedTarget(opponent.getId()));
-                game.addEffect(effect, source);
-                return true;
-            }
-        }
-        return false;
-    }
-}
-
-class AttacksIfAbleTargetPlayerSourceEffect extends RequirementEffect {
-
-    public AttacksIfAbleTargetPlayerSourceEffect() {
-        super(Duration.EndOfTurn);
-        staticText = "{this} attacks that player this combat if able";
-    }
-
-    public AttacksIfAbleTargetPlayerSourceEffect(final AttacksIfAbleTargetPlayerSourceEffect effect) {
-        super(effect);
-    }
-
-    @Override
-    public AttacksIfAbleTargetPlayerSourceEffect copy() {
-        return new AttacksIfAbleTargetPlayerSourceEffect(this);
-    }
-
-    @Override
-    public boolean applies(Permanent permanent, Ability source, Game game) {
-        if (permanent.getId().equals(source.getSourceId())) {
-            return true;
-        }
-        return false;
-    }
-
-    @Override
-    public boolean mustAttack(Game game) {
-        return true;
-    }
-
-    @Override
-    public boolean mustBlock(Game game) {
-        return false;
-    }
-
-    @Override
-    public UUID mustAttackDefender(Ability source, Game game) {
-        return getTargetPointer().getFirst(game, source);
-    }
-
-}
